removing ensure_iface_setup
ensure_iface_setup was called from send_ra.  The responsibility of setting
up ifaces is not in the send function.  If an interface is not setup, it
should not be scheduled and netlink should set it up when the kernel says
the interface is ready.  However, I suppose if we somehow get a solicit on
the interface and the interface is not ready, we should attempt to make it
ready at that point.