send.c: Fix by Scott Shambarger for bug 1048653
This code was taken off bugzilla.redhat.com.

Bug 1048653 - UnicastOnly option still sends multicast RA

Updated send.c to not call send_ra() from send_ra_forall() when
dest==NULL and UnicastOnly is set.  This appears to be the only
hole in the logic I can find, but still resets the next_multicast
timer to avoid any poll timer issues.

Tested, and it works as expected (UnicastOnly on and off).