add a syscalls layer
To do testing, separate out the important system calls.  The system
calls will be reimplemented in a test layer to drive radvd during
tests.  Two radvd executibles will be created, radvd, linked to
a systemcall passthrough layer, and radvd_test, which will be linked
to the syscalls test layer.

Hopefully when all this is done, radvd can be tested by non root and
will be able to run on the syscalls layer not knowing, or caring if
it runs on the actual OS calls or the fake, test systemcalls.