device-common.c: move check_ip6_forwarding here
check_ip6_forwarding was defined in radvd.c, but that's not a good place
for it to live.  It's better in device-common.c.

More function will eventually have to be migrated out of radvd.c, or at
lease the main function in radvd.c should be moved out to enable check
programs to link.